Time operations in a pipeline Measure how long a data operation or function takes within a pipeline (|>). This can be used to check performance of steps in your data workflow.

Time operations in a pipeline Measure how long a data operation or function takes within a pipeline (|>). This can be used to check performance of steps in your data workflow.

time_pipe(
  .data,
  action = NULL,
  label = NULL,
  log_file = NULL,
  console = TRUE,
  time_unit = c("secs", "millisecs", "mins")
)

Arguments

.data

A data object to pass through the pipeline.

action

Optional. A function call to apply to .data. If not provided, the function will time .data itself.

label

Optional. A descriptive name for the operation. If not provided, a name will be generated automatically.

log_file

Optional. File path to write timing logs. If NULL, messages are only printed to the console.

console

Logical. Whether to print timing messages to the console. Default is TRUE.

time_unit

Character. Unit of time to report. One of "secs", "millisecs", or "mins".

Value

Returns the result of the pipeline step, unchanged. Timing messages are printed or logged separately.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
mtcars |>
  time_pipe(group_by(cyl), "grouping") |>
  time_pipe(summarise(avg_hp = mean(hp)), "aggregation")
#> [2025-08-14 19:16:37.097] grouping: 0.0031 secs elapsed
#> [2025-08-14 19:16:37.100] aggregation: 0.0048 secs elapsed
#> # A tibble: 3 × 2
#>     cyl avg_hp
#>   <dbl>  <dbl>
#> 1     4   82.6
#> 2     6  122. 
#> 3     8  209. 

mtcars |>
  mutate(hp2 = {
    Sys.sleep(1)
    hp * 2
  }) |>
  time_pipe("entire pipeline")
#> [2025-08-14 19:16:37.121] entire pipeline: 1.0032 secs elapsed
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb hp2
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 220
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 220
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1 186
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 220
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2 350
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1 210
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4 490
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2 124
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2 190
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 246
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 246
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3 360
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3 360
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3 360
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4 410
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4 430
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4 460
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1 132
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 104
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1 130
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1 194
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2 300
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2 300
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4 490
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2 350
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1 132
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2 182
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 226
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4 528
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6 350
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8 670
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2 218